==> 03_regression.Rmd <==
---
title: "Regression Analysis"
author: "Ian Curtis"
date: "2022-11-21"
output: pdf_document
---

```{r}
library(tidyverse)
library(caret)
library(kableExtra)
library(webshot)
```

```{r}
train <- read_csv('data/train_data.csv')
test <- read_csv('data/test_data.csv')
```

```{r}
train_reg <- train %>% 
  select(-c(pitch_number, pitch_type))
test_reg <- test %>% 
  select(-c(pitch_number, pitch_type))
```

```{r}
preds <- function(data, reg_model, model_name) {
  data <- data %>% mutate(
  yhat_reg = as_vector(reg_model['predicted']),
  pred_dist_reg = as.factor(case_when(
          (yhat_reg <= 70) ~ 'Zone 1',
          (yhat_reg > 70 & yhat_reg <= 140) ~ 'Zone 2',
          (yhat_reg > 140 & yhat_reg <= 210) ~ 'Zone 3',
          (yhat_reg > 210 & yhat_reg <= 280) ~ 'Zone 4',
          (yhat_reg > 280 & yhat_reg <= 350) ~ 'Zone 5',
          (yhat_reg > 350) ~ 'Zone 6',
        )),
  )
  
  cm <- confusionMatrix(data$pred_dist_reg, as.factor(data$dist_categ))
  
  return(
    list(model = model_name, 
         mtx = cm$table, 
         accuracy = cm$overall[['Accuracy']]
         ))
  
}
```


# Regression Analysis

```{r}
olsr <- function(model_name, data) {
  
  # Y vector of all dependent variable values
  Y <- data$hit_distance

  # X vector of all independent variable terms (with spot for intercept)
  intercept <- rep(1, nrow(data)) # large vector of ones
  almostX <- data %>% select(-c(hit_distance, dist_categ)) # take out the dependent variable
  X <- as.matrix(cbind(intercept, almostX)) # put intercept vector at beginning

  # X'X matrix
  Xt <- t(X) # transpose of the X independent values
  XtX <- Xt %*% X # multiply the X and its transpose

  XtXi <- solve(XtX)

  # X'y matrix
  XtY <- Xt %*% Y

  # Beta parameter estimations for the least squares line
  B <- XtXi %*% XtY

  # Predicted values using the generated beta estimations
  y_hat <- X %*% B

  # Residuals (error)
  e <- Y - y_hat

  # Sum of Squares
  sst <- (t(Y) %*% Y) - (nrow(data) * mean(Y)**2)
  sse <- t(e) %*% e
  ssr <- sst - sse
  msr <- ssr / (ncol(data) - 2) # minus 2 because dist_categ is not part of analysis
  mse <- sse / (nrow(data) - (ncol(data) - 1)) # minus 1 because dist_categ is not part of analysis
  
  # R squared
  r2 <- ssr / sst
  
  s2 <- sse / (nrow(data) - (ncol(data) - 1))
  
  vcov <- matrix(unlist(drop(s2) * XtXi), ncol = (ncol(data) - 1))
  stderr <- sqrt(diag(vcov))
  t <- B / stderr
  p_t <- dt(t, df = nrow(data) - (ncol(data) - 1))
  p_t <- lapply(list(p_t), round, 4)

  f <- msr / mse
  p_overall <- round(pf(f, (ncol(data) - 1) - 1, nrow(data) - (ncol(data) - 1), lower.tail = F), 4)
  
  results <- list('model' = model_name, 'betas' = B, 'predicted' = y_hat, 'errors' = e, 'sst' = sst, 'sse' = sse, 'ssr'= ssr, 'msr' = msr, 'mse' = mse, 'r2' = r2, 's2' = s2, 'std_errs' = stderr, 't' = t, 'p_t' = p_t, 'f' = f, 'p_overall' = p_overall)
  
  return(results)
}

evaluate_reg <- function(reg_output) {
  toadd <- c(reg_output['model'], reg_output['sse'], reg_output['mse'], reg_output['r2'], reg_output['f'], reg_output['p_overall'])
  
  return(toadd)
}

```

## Full Model

```{r}
full_model <- olsr('full', train_reg)
full_model['p_t']
# preds(train_reg, full_model, 'full_model')
```


```{r}
RM1 <- train_reg %>% 
  select(-c(balls, strikes, outs_when_up, inning, month, d_stand_l, d_pthrow_l, i_release_balls, i_release_inn, i_release_spin, i_ball_strike, i_spin_month))

reduced1 <- olsr('RM1', RM1)
reduced1['p_t']

# preds(train_reg, reduced1, 'reduced1')
```

```{r}
RM2 <- RM1 %>% 
  select(-c(i_release_outs, i_release_SI, i_release_FC, i_release_FF, i_release_CU, i_release_SL, i_release_CH, d_inning_top), -starts_with('d_pitch'))

reduced2 <- olsr('RM2', RM2)
# reduced2['p_t']

preds(train_reg, reduced2, 'reduced2')
```

```{r}
RM3 <- train_reg %>% 
  select(-c(i_release_SI, i_release_FC, i_release_FF, i_release_CU, i_release_SL, i_release_CH, balls, strikes, outs_when_up, inning, month), -starts_with('d_pitch'))
reduced3 <- olsr('RM3', RM3)

reduced3['p_t']

# preds(train_reg, reduced3, 'reduced3')
```

```{r}
RM4 <- RM3 %>% 
  select(-c(d_stand_l, i_release_outs, i_release_inn, i_spin_month, d_inning_top, release_spin_rate))

reduced4 <- olsr('RM4', RM4)
# reduced4['p_t']

preds(train_reg, reduced4, 'reduced4')
```

```{r}
RM5 <- train_reg %>% select(hit_distance, dist_categ, launch_angle, launch_speed, release_spin_rate, sqrt_pitch_num, i_release_launcha, i_release_launchs, i_launches, i_launcha_spin)

reduced5 <- olsr('RM5', RM5)
# reduced5['p_t']

preds(train_reg, reduced5, 'reduced5')
```

```{r}
eval_reg <- bind_rows(evaluate_reg(full_model), evaluate_reg(reduced1), evaluate_reg(reduced2), evaluate_reg(reduced3), evaluate_reg(reduced4), evaluate_reg(reduced5))

colnames(eval_reg) <- c('model', 'sse', 'mse', 'r2', 'f_overall', 'p_overall')
```

```{r resid-plot-red5}
resid_plot <- as_tibble(c(reduced5['errors'], reduced5['predicted'])) %>% 
  ggplot(aes(x = predicted, y = errors)) +
  geom_point()
ggsave('resid_plot.pdf', path = './figures')
```

Assumptions are not met so analysis will not continue for linear regression.
